requirements.txt
streamlit
pandas
matplotlib
plotly
scikit-learn
prophet
openai
python-dotenv
